+++ src/Presentation/Nop.Web/Factories/CatalogModelFactory.cs
@@ -1,0 +1,1721 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Diagnostics;
+using System.Linq;
+using System.Web;
+using System.Web.Mvc;
+using Nop.Core;
+using Nop.Core.Caching;
+using Nop.Core.Domain.Blogs;
+using Nop.Core.Domain.Catalog;
+using Nop.Core.Domain.Common;
+using Nop.Core.Domain.Forums;
+using Nop.Core.Domain.Media;
+using Nop.Core.Domain.Vendors;
+using Nop.Services.Catalog;
+using Nop.Services.Common;
+using Nop.Services.Customers;
+using Nop.Services.Directory;
+using Nop.Services.Events;
+using Nop.Services.Localization;
+using Nop.Services.Media;
+using Nop.Services.Security;
+using Nop.Services.Seo;
+using Nop.Services.Stores;
+using Nop.Services.Tax;
+using Nop.Services.Topics;
+using Nop.Services.Vendors;
+using Nop.Web.Extensions;
+using Nop.Web.Framework.Events;
+using Nop.Web.Infrastructure.Cache;
+using Nop.Web.Models.Catalog;
+using Nop.Web.Models.Media;
+
+namespace Nop.Web.Factories
+{
+    public partial class CatalogModelFactory : ICatalogModelFactory
+    {
+		#region Fields
+
+        private readonly ICategoryService _categoryService;
+        private readonly IManufacturerService _manufacturerService;
+        private readonly IProductService _productService;
+        private readonly IVendorService _vendorService;
+        private readonly ICategoryTemplateService _categoryTemplateService;
+        private readonly IManufacturerTemplateService _manufacturerTemplateService;
+        private readonly IWorkContext _workContext;
+        private readonly IStoreContext _storeContext;
+        private readonly ITaxService _taxService;
+        private readonly ICurrencyService _currencyService;
+        private readonly IPictureService _pictureService;
+        private readonly ILocalizationService _localizationService;
+        private readonly IPriceCalculationService _priceCalculationService;
+        private readonly IPriceFormatter _priceFormatter;
+        private readonly IWebHelper _webHelper;
+        private readonly ISpecificationAttributeService _specificationAttributeService;
+        private readonly IProductTagService _productTagService;
+        private readonly IAclService _aclService;
+        private readonly IStoreMappingService _storeMappingService;
+        private readonly IPermissionService _permissionService;
+        private readonly ITopicService _topicService;
+        private readonly IEventPublisher _eventPublisher;
+        private readonly ISearchTermService _searchTermService;
+        private readonly IMeasureService _measureService;
+        private readonly HttpContextBase _httpContext;
+        private readonly MediaSettings _mediaSettings;
+        private readonly CatalogSettings _catalogSettings;
+        private readonly VendorSettings _vendorSettings;
+        private readonly BlogSettings _blogSettings;
+        private readonly ForumSettings _forumSettings;
+        private readonly ICacheManager _cacheManager;
+
+        #endregion
+
+        #region Constructors
+
+        public CatalogModelFactory(ICategoryService categoryService, 
+            IManufacturerService manufacturerService,
+            IProductService productService, 
+            IVendorService vendorService,
+            ICategoryTemplateService categoryTemplateService,
+            IManufacturerTemplateService manufacturerTemplateService,
+            IWorkContext workContext, 
+            IStoreContext storeContext,
+            ITaxService taxService, 
+            ICurrencyService currencyService,
+            IPictureService pictureService, 
+            ILocalizationService localizationService,
+            IPriceCalculationService priceCalculationService,
+            IPriceFormatter priceFormatter,
+            IWebHelper webHelper, 
+            ISpecificationAttributeService specificationAttributeService,
+            IProductTagService productTagService,
+            IAclService aclService,
+            IStoreMappingService storeMappingService,
+            IPermissionService permissionService, 
+            ITopicService topicService,
+            IEventPublisher eventPublisher,
+            ISearchTermService searchTermService,
+            IMeasureService measureService,
+            HttpContextBase httpContext,
+            MediaSettings mediaSettings,
+            CatalogSettings catalogSettings,
+            VendorSettings vendorSettings,
+            BlogSettings blogSettings,
+            ForumSettings  forumSettings,
+            ICacheManager cacheManager)
+        {
+            this._categoryService = categoryService;
+            this._manufacturerService = manufacturerService;
+            this._productService = productService;
+            this._vendorService = vendorService;
+            this._categoryTemplateService = categoryTemplateService;
+            this._manufacturerTemplateService = manufacturerTemplateService;
+            this._workContext = workContext;
+            this._storeContext = storeContext;
+            this._taxService = taxService;
+            this._currencyService = currencyService;
+            this._pictureService = pictureService;
+            this._localizationService = localizationService;
+            this._priceCalculationService = priceCalculationService;
+            this._priceFormatter = priceFormatter;
+            this._webHelper = webHelper;
+            this._specificationAttributeService = specificationAttributeService;
+            this._productTagService = productTagService;
+            this._aclService = aclService;
+            this._storeMappingService = storeMappingService;
+            this._permissionService = permissionService;
+            this._topicService = topicService;
+            this._eventPublisher = eventPublisher;
+            this._searchTermService = searchTermService;
+            this._measureService = measureService;
+            this._httpContext = httpContext;
+            this._mediaSettings = mediaSettings;
+            this._catalogSettings = catalogSettings;
+            this._vendorSettings = vendorSettings;
+            this._blogSettings = blogSettings;
+            this._forumSettings = forumSettings;
+            this._cacheManager = cacheManager;
+        }
+
+        #endregion
+
+        #region Utilities
+        
+        protected virtual void PrepareSortingOptions(CatalogPagingFilteringModel pagingFilteringModel, CatalogPagingFilteringModel command)
+        {
+            if (pagingFilteringModel == null)
+                throw new ArgumentNullException("pagingFilteringModel");
+
+            if (command == null)
+                throw new ArgumentNullException("command");
+
+            var allDisabled = _catalogSettings.ProductSortingEnumDisabled.Count == Enum.GetValues(typeof(ProductSortingEnum)).Length;
+            pagingFilteringModel.AllowProductSorting = _catalogSettings.AllowProductSorting && !allDisabled;
+
+            var activeOptions = Enum.GetValues(typeof(ProductSortingEnum)).Cast<int>()
+                .Except(_catalogSettings.ProductSortingEnumDisabled)
+                .Select((idOption) =>
+                {
+                    int order;
+                    return new KeyValuePair<int, int>(idOption, _catalogSettings.ProductSortingEnumDisplayOrder.TryGetValue(idOption, out order) ? order : idOption);
+                })
+                .OrderBy(x => x.Value);
+            if (command.OrderBy == null)
+                command.OrderBy = allDisabled ? 0 : activeOptions.First().Key;
+
+            if (pagingFilteringModel.AllowProductSorting)
+            {
+                foreach (var option in activeOptions)
+                {
+                    var currentPageUrl = _webHelper.GetThisPageUrl(true);
+                    var sortUrl = _webHelper.ModifyQueryString(currentPageUrl, "orderby=" + (option.Key).ToString(), null);
+
+                    var sortValue = ((ProductSortingEnum)option.Key).GetLocalizedEnum(_localizationService, _workContext);
+                    pagingFilteringModel.AvailableSortOptions.Add(new SelectListItem
+                    {
+                        Text = sortValue,
+                        Value = sortUrl,
+                        Selected = option.Key == command.OrderBy
+                    });
+                }
+            }
+        }
+        
+        protected virtual void PrepareViewModes(CatalogPagingFilteringModel pagingFilteringModel, CatalogPagingFilteringModel command)
+        {
+            if (pagingFilteringModel == null)
+                throw new ArgumentNullException("pagingFilteringModel");
+
+            if (command == null)
+                throw new ArgumentNullException("command");
+
+            pagingFilteringModel.AllowProductViewModeChanging = _catalogSettings.AllowProductViewModeChanging;
+
+            var viewMode = !string.IsNullOrEmpty(command.ViewMode)
+                ? command.ViewMode
+                : _catalogSettings.DefaultViewMode;
+            pagingFilteringModel.ViewMode = viewMode;
+            if (pagingFilteringModel.AllowProductViewModeChanging)
+            {
+                var currentPageUrl = _webHelper.GetThisPageUrl(true);
+                //grid
+                pagingFilteringModel.AvailableViewModes.Add(new SelectListItem
+                {
+                    Text = _localizationService.GetResource("Catalog.ViewMode.Grid"),
+                    Value = _webHelper.ModifyQueryString(currentPageUrl, "viewmode=grid", null),
+                    Selected = viewMode == "grid"
+                });
+                //list
+                pagingFilteringModel.AvailableViewModes.Add(new SelectListItem
+                {
+                    Text = _localizationService.GetResource("Catalog.ViewMode.List"),
+                    Value = _webHelper.ModifyQueryString(currentPageUrl, "viewmode=list", null),
+                    Selected = viewMode == "list"
+                });
+            }
+
+        }
+        
+        protected virtual void PreparePageSizeOptions(CatalogPagingFilteringModel pagingFilteringModel, CatalogPagingFilteringModel command,
+            bool allowCustomersToSelectPageSize, string pageSizeOptions, int fixedPageSize)
+        {
+            if (pagingFilteringModel == null)
+                throw new ArgumentNullException("pagingFilteringModel");
+
+            if (command == null)
+                throw new ArgumentNullException("command");
+
+            if (command.PageNumber <= 0)
+            {
+                command.PageNumber = 1;
+            }
+            pagingFilteringModel.AllowCustomersToSelectPageSize = false;
+            if (allowCustomersToSelectPageSize && pageSizeOptions != null)
+            {
+                var pageSizes = pageSizeOptions.Split(new [] { ',', ' ' }, StringSplitOptions.RemoveEmptyEntries);
+
+                if (pageSizes.Any())
+                {
+                    // get the first page size entry to use as the default (category page load) or if customer enters invalid value via query string
+                    if (command.PageSize <= 0 || !pageSizes.Contains(command.PageSize.ToString()))
+                    {
+                        int temp;
+                        if (int.TryParse(pageSizes.FirstOrDefault(), out temp))
+                        {
+                            if (temp > 0)
+                            {
+                                command.PageSize = temp;
+                            }
+                        }
+                    }
+
+                    var currentPageUrl = _webHelper.GetThisPageUrl(true);
+                    var sortUrl = _webHelper.ModifyQueryString(currentPageUrl, "pagesize={0}", null);
+                    sortUrl = _webHelper.RemoveQueryString(sortUrl, "pagenumber");
+
+                    foreach (var pageSize in pageSizes)
+                    {
+                        int temp;
+                        if (!int.TryParse(pageSize, out temp))
+                        {
+                            continue;
+                        }
+                        if (temp <= 0)
+                        {
+                            continue;
+                        }
+
+                        pagingFilteringModel.PageSizeOptions.Add(new SelectListItem
+                        {
+                            Text = pageSize,
+                            Value = String.Format(sortUrl, pageSize),
+                            Selected = pageSize.Equals(command.PageSize.ToString(), StringComparison.InvariantCultureIgnoreCase)
+                        });
+                    }
+
+                    if (pagingFilteringModel.PageSizeOptions.Any())
+                    {
+                        pagingFilteringModel.PageSizeOptions = pagingFilteringModel.PageSizeOptions.OrderBy(x => int.Parse(x.Text)).ToList();
+                        pagingFilteringModel.AllowCustomersToSelectPageSize = true;
+
+                        if (command.PageSize <= 0)
+                        {
+                            command.PageSize = int.Parse(pagingFilteringModel.PageSizeOptions.FirstOrDefault().Text);
+                        }
+                    }
+                }
+            }
+            else
+            {
+                //customer is not allowed to select a page size
+                command.PageSize = fixedPageSize;
+            }
+
+            //ensure pge size is specified
+            if (command.PageSize <= 0)
+            {
+                command.PageSize = fixedPageSize;
+            }
+        }
+        
+        protected virtual List<int> GetChildCategoryIds(int parentCategoryId)
+        {
+            string cacheKey = string.Format(ModelCacheEventConsumer.CATEGORY_CHILD_IDENTIFIERS_MODEL_KEY, 
+                parentCategoryId, 
+                string.Join(",", _workContext.CurrentCustomer.GetCustomerRoleIds()), 
+                _storeContext.CurrentStore.Id);
+            return _cacheManager.Get(cacheKey, () =>
+            {
+                var categoriesIds = new List<int>();
+                var categories = _categoryService.GetAllCategoriesByParentCategoryId(parentCategoryId);
+                foreach (var category in categories)
+                {
+                    categoriesIds.Add(category.Id);
+                    categoriesIds.AddRange(GetChildCategoryIds(category.Id));
+                }
+                return categoriesIds;
+            });
+        }
+
+        /// <summary>
+        /// Prepare category (simple) models
+        /// </summary>
+        /// <param name="rootCategoryId">Root category identifier</param>
+        /// <param name="loadSubCategories">A value indicating whether subcategories should be loaded</param>
+        /// <param name="allCategories">All available categories; pass null to load them internally</param>
+        /// <returns>Category models</returns>
+        protected virtual IList<CategorySimpleModel> PrepareCategorySimpleModels(int rootCategoryId,
+            bool loadSubCategories = true,IList<Category> allCategories = null)
+        {
+            var result = new List<CategorySimpleModel>();
+
+            //little hack for performance optimization.
+            //we know that this method is used to load top and left menu for categories.
+            //it'll load all categories anyway.
+            //so there's no need to invoke "GetAllCategoriesByParentCategoryId" multiple times (extra SQL commands) to load childs
+            //so we load all categories at once
+            //if you don't like this implementation if you can uncomment the line below (old behavior) and comment several next lines (before foreach)
+            //var categories = _categoryService.GetAllCategoriesByParentCategoryId(rootCategoryId);
+            if (allCategories == null)
+            {
+                //load categories if null passed
+                //we implemeneted it this way for performance optimization - recursive iterations (below)
+                //this way all categories are loaded only once
+                allCategories = _categoryService.GetAllCategories(storeId: _storeContext.CurrentStore.Id);
+            }
+            var categories = allCategories.Where(c => c.ParentCategoryId == rootCategoryId).ToList();
+            foreach (var category in categories)
+            {
+                var categoryModel = new CategorySimpleModel
+                {
+                    Id = category.Id,
+                    Name = category.GetLocalized(x => x.Name),
+                    SeName = category.GetSeName(),
+                    IncludeInTopMenu = category.IncludeInTopMenu
+                };
+
+                //number of products in each category
+                if (_catalogSettings.ShowCategoryProductNumber)
+                {
+                    string cacheKey = string.Format(ModelCacheEventConsumer.CATEGORY_NUMBER_OF_PRODUCTS_MODEL_KEY,
+                        string.Join(",", _workContext.CurrentCustomer.GetCustomerRoleIds()),
+                        _storeContext.CurrentStore.Id, 
+                        category.Id);
+                    categoryModel.NumberOfProducts = _cacheManager.Get(cacheKey, () =>
+                    {
+                        var categoryIds = new List<int>();
+                        categoryIds.Add(category.Id);
+                        //include subcategories
+                        if (_catalogSettings.ShowCategoryProductNumberIncludingSubcategories)
+                            categoryIds.AddRange(GetChildCategoryIds(category.Id));
+                        return _productService.GetNumberOfProductsInCategory(categoryIds, _storeContext.CurrentStore.Id);
+                    });
+                }
+
+                if (loadSubCategories)
+                {
+                    var subCategories = PrepareCategorySimpleModels(category.Id, loadSubCategories, allCategories);
+                    categoryModel.SubCategories.AddRange(subCategories);
+                }
+                result.Add(categoryModel);
+            }
+
+            return result;
+        }
+
+        #endregion
+
+        #region Products
+
+        public virtual ProductReviewOverviewModel PrepareProductReviewOverviewModel(Product product)
+        {
+            ProductReviewOverviewModel productReview;
+
+            if (_catalogSettings.ShowProductReviewsPerStore)
+            {
+                string cacheKey = string.Format(ModelCacheEventConsumer.PRODUCT_REVIEWS_MODEL_KEY, product.Id, _storeContext.CurrentStore.Id);
+
+                productReview = _cacheManager.Get(cacheKey, () =>
+                {
+                    return new ProductReviewOverviewModel
+                    {
+                        RatingSum = product.ProductReviews
+                                .Where(pr => pr.IsApproved && pr.StoreId == _storeContext.CurrentStore.Id)
+                                .Sum(pr => pr.Rating),
+                        TotalReviews = product
+                                .ProductReviews
+                                .Count(pr => pr.IsApproved && pr.StoreId == _storeContext.CurrentStore.Id)
+                    };
+                });
+            }
+            else
+            {
+                productReview = new ProductReviewOverviewModel()
+                {
+                    RatingSum = product.ApprovedRatingSum,
+                    TotalReviews = product.ApprovedTotalReviews
+                };
+            }
+            if (productReview != null)
+            {
+                productReview.ProductId = product.Id;
+                productReview.AllowCustomerReviews = product.AllowCustomerReviews;
+            }
+            return productReview;
+        }
+
+        public virtual IEnumerable<ProductOverviewModel> PrepareProductOverviewModels(IEnumerable<Product> products, 
+            bool preparePriceModel = true, bool preparePictureModel = true,
+            int? productThumbPictureSize = null, bool prepareSpecificationAttributes = false,
+            bool forceRedirectionAfterAddingToCart = false)
+        {
+            if (products == null)
+                throw new ArgumentNullException("products");
+
+            var models = new List<ProductOverviewModel>();
+            foreach (var product in products)
+            {
+                var model = new ProductOverviewModel
+                {
+                    Id = product.Id,
+                    Name = product.GetLocalized(x => x.Name),
+                    ShortDescription = product.GetLocalized(x => x.ShortDescription),
+                    FullDescription = product.GetLocalized(x => x.FullDescription),
+                    SeName = product.GetSeName(),
+                    Sku = product.Sku,
+                    ProductType = product.ProductType,
+                    MarkAsNew = product.MarkAsNew &&
+                        (!product.MarkAsNewStartDateTimeUtc.HasValue || product.MarkAsNewStartDateTimeUtc.Value < DateTime.UtcNow) &&
+                        (!product.MarkAsNewEndDateTimeUtc.HasValue || product.MarkAsNewEndDateTimeUtc.Value > DateTime.UtcNow)
+                };
+                //price
+                if (preparePriceModel)
+                {
+                    #region Prepare product price
+
+                    var priceModel = new ProductOverviewModel.ProductPriceModel
+                    {
+                        ForceRedirectionAfterAddingToCart = forceRedirectionAfterAddingToCart
+                    };
+
+                    switch (product.ProductType)
+                    {
+                        case ProductType.GroupedProduct:
+                            {
+                                #region Grouped product
+
+                                var associatedProducts = _productService.GetAssociatedProducts(product.Id, _storeContext.CurrentStore.Id);
+
+                                //add to cart button (ignore "DisableBuyButton" property for grouped products)
+                                priceModel.DisableBuyButton = !_permissionService.Authorize(StandardPermissionProvider.EnableShoppingCart) ||
+                                    !_permissionService.Authorize(StandardPermissionProvider.DisplayPrices);
+
+                                //add to wishlist button (ignore "DisableWishlistButton" property for grouped products)
+                                priceModel.DisableWishlistButton = !_permissionService.Authorize(StandardPermissionProvider.EnableWishlist) ||
+                                    !_permissionService.Authorize(StandardPermissionProvider.DisplayPrices);
+
+                                //compare products
+                                priceModel.DisableAddToCompareListButton = !_catalogSettings.CompareProductsEnabled;
+                                switch (associatedProducts.Count)
+                                {
+                                    case 0:
+                                        {
+                                            //no associated products
+                                        }
+                                        break;
+                                    default:
+                                        {
+                                            //we have at least one associated product
+                                            //compare products
+                                            priceModel.DisableAddToCompareListButton = !_catalogSettings.CompareProductsEnabled;
+                                            //priceModel.AvailableForPreOrder = false;
+
+                                            if (_permissionService.Authorize(StandardPermissionProvider.DisplayPrices))
+                                            {
+                                                //find a minimum possible price
+                                                decimal? minPossiblePrice = null;
+                                                Product minPriceProduct = null;
+                                                foreach (var associatedProduct in associatedProducts)
+                                                {
+                                                    //calculate for the maximum quantity (in case if we have tier prices)
+                                                    var tmpPrice = _priceCalculationService.GetFinalPrice(associatedProduct,
+                                                        _workContext.CurrentCustomer, decimal.Zero, true, int.MaxValue);
+                                                    if (!minPossiblePrice.HasValue || tmpPrice < minPossiblePrice.Value)
+                                                    {
+                                                        minPriceProduct = associatedProduct;
+                                                        minPossiblePrice = tmpPrice;
+                                                    }
+                                                }
+                                                if (minPriceProduct != null && !minPriceProduct.CustomerEntersPrice)
+                                                {
+                                                    if (minPriceProduct.CallForPrice)
+                                                    {
+                                                        priceModel.OldPrice = null;
+                                                        priceModel.Price = _localizationService.GetResource("Products.CallForPrice");
+                                                    }
+                                                    else if (minPossiblePrice.HasValue)
+                                                    {
+                                                        //calculate prices
+                                                        decimal taxRate;
+                                                        decimal finalPriceBase = _taxService.GetProductPrice(minPriceProduct, minPossiblePrice.Value, out taxRate);
+                                                        decimal finalPrice = _currencyService.ConvertFromPrimaryStoreCurrency(finalPriceBase, _workContext.WorkingCurrency);
+
+                                                        priceModel.OldPrice = null;
+                                                        priceModel.Price = String.Format(_localizationService.GetResource("Products.PriceRangeFrom"), _priceFormatter.FormatPrice(finalPrice));
+                                                        priceModel.PriceValue = finalPrice;
+
+                                                        //PAngV baseprice (used in Germany)
+                                                        priceModel.BasePricePAngV = product.FormatBasePrice(finalPrice,
+                                                            _localizationService, _measureService, _currencyService, _workContext, _priceFormatter);
+                                                    }
+                                                    else
+                                                    {
+                                                        //Actually it's not possible (we presume that minimalPrice always has a value)
+                                                        //We never should get here
+                                                        Debug.WriteLine("Cannot calculate minPrice for product #{0}", product.Id);
+                                                    }
+                                                }
+                                            }
+                                            else
+                                            {
+                                                //hide prices
+                                                priceModel.OldPrice = null;
+                                                priceModel.Price = null;
+                                            }
+                                        }
+                                        break;
+                                }
+
+                                #endregion
+                            }
+                            break;
+                        case ProductType.SimpleProduct:
+                        default:
+                            {
+                                #region Simple product
+
+                                //add to cart button
+                                priceModel.DisableBuyButton = product.DisableBuyButton ||
+                                    !_permissionService.Authorize(StandardPermissionProvider.EnableShoppingCart) ||
+                                    !_permissionService.Authorize(StandardPermissionProvider.DisplayPrices);
+
+                                //add to wishlist button
+                                priceModel.DisableWishlistButton = product.DisableWishlistButton ||
+                                    !_permissionService.Authorize(StandardPermissionProvider.EnableWishlist) ||
+                                    !_permissionService.Authorize(StandardPermissionProvider.DisplayPrices);
+                                //compare products
+                                priceModel.DisableAddToCompareListButton = !_catalogSettings.CompareProductsEnabled;
+
+                                //rental
+                                priceModel.IsRental = product.IsRental;
+
+                                //pre-order
+                                if (product.AvailableForPreOrder)
+                                {
+                                    priceModel.AvailableForPreOrder = !product.PreOrderAvailabilityStartDateTimeUtc.HasValue ||
+                                        product.PreOrderAvailabilityStartDateTimeUtc.Value >= DateTime.UtcNow;
+                                    priceModel.PreOrderAvailabilityStartDateTimeUtc = product.PreOrderAvailabilityStartDateTimeUtc;
+                                }
+
+                                //prices
+                                if (_permissionService.Authorize(StandardPermissionProvider.DisplayPrices))
+                                {
+                                    if (!product.CustomerEntersPrice)
+                                    {
+                                        if (product.CallForPrice)
+                                        {
+                                            //call for price
+                                            priceModel.OldPrice = null;
+                                            priceModel.Price = _localizationService.GetResource("Products.CallForPrice");
+                                        }
+                                        else
+                                        {
+                                            //prices
+
+                                            //calculate for the maximum quantity (in case if we have tier prices)
+                                            decimal minPossiblePrice = _priceCalculationService.GetFinalPrice(product,
+                                                _workContext.CurrentCustomer, decimal.Zero, true, int.MaxValue);
+
+                                            decimal taxRate;
+                                            decimal oldPriceBase = _taxService.GetProductPrice(product, product.OldPrice, out taxRate);
+                                            decimal finalPriceBase = _taxService.GetProductPrice(product, minPossiblePrice, out taxRate);
+
+                                            decimal oldPrice = _currencyService.ConvertFromPrimaryStoreCurrency(oldPriceBase, _workContext.WorkingCurrency);
+                                            decimal finalPrice = _currencyService.ConvertFromPrimaryStoreCurrency(finalPriceBase, _workContext.WorkingCurrency);
+
+                                            //do we have tier prices configured?
+                                            var tierPrices = new List<TierPrice>();
+                                            if (product.HasTierPrices)
+                                            {
+                                                tierPrices.AddRange(product.TierPrices.OrderBy(tp => tp.Quantity)
+                                                    .FilterByStore(_storeContext.CurrentStore.Id)
+                                                    .FilterForCustomer(_workContext.CurrentCustomer)
+                                                    .FilterByDate()
+                                                    .RemoveDuplicatedQuantities());
+                                            }
+                                            //When there is just one tier price (with  qty 1), there are no actual savings in the list.
+                                            var displayFromMessage = tierPrices.Any() && !(tierPrices.Count == 1 && tierPrices[0].Quantity <= 1);
+                                            if (displayFromMessage)
+                                            {
+                                                priceModel.OldPrice = null;
+                                                priceModel.Price = String.Format(_localizationService.GetResource("Products.PriceRangeFrom"), _priceFormatter.FormatPrice(finalPrice));
+                                                priceModel.PriceValue = finalPrice;
+                                            }
+                                            else
+                                            {
+                                                if (finalPriceBase != oldPriceBase && oldPriceBase != decimal.Zero)
+                                                {
+                                                    priceModel.OldPrice = _priceFormatter.FormatPrice(oldPrice);
+                                                    priceModel.Price = _priceFormatter.FormatPrice(finalPrice);
+                                                    priceModel.PriceValue = finalPrice;
+                                                }
+                                                else
+                                                {
+                                                    priceModel.OldPrice = null;
+                                                    priceModel.Price = _priceFormatter.FormatPrice(finalPrice);
+                                                    priceModel.PriceValue = finalPrice;
+                                                }
+                                            }
+                                            if (product.IsRental)
+                                            {
+                                                //rental product
+                                                priceModel.OldPrice = _priceFormatter.FormatRentalProductPeriod(product, priceModel.OldPrice);
+                                                priceModel.Price = _priceFormatter.FormatRentalProductPeriod(product, priceModel.Price);
+                                            }
+
+
+                                            //property for German market
+                                            //we display tax/shipping info only with "shipping enabled" for this product
+                                            //we also ensure this it's not free shipping
+                                            priceModel.DisplayTaxShippingInfo = _catalogSettings.DisplayTaxShippingInfoProductBoxes
+                                                && product.IsShipEnabled &&
+                                                !product.IsFreeShipping;
+
+
+                                            //PAngV baseprice (used in Germany)
+                                            priceModel.BasePricePAngV = product.FormatBasePrice(finalPrice,
+                                                _localizationService, _measureService, _currencyService, _workContext, _priceFormatter);
+                                        }
+                                    }
+                                }
+                                else
+                                {
+                                    //hide prices
+                                    priceModel.OldPrice = null;
+                                    priceModel.Price = null;
+                                }
+
+                                #endregion
+                            }
+                            break;
+                    }
+
+                    model.ProductPrice = priceModel;
+
+                    #endregion
+                }
+
+                //picture
+                if (preparePictureModel)
+                {
+                    #region Prepare product picture
+
+                    //If a size has been set in the view, we use it in priority
+                    int pictureSize = productThumbPictureSize.HasValue ? productThumbPictureSize.Value : _mediaSettings.ProductThumbPictureSize;
+                    //prepare picture model
+                    var defaultProductPictureCacheKey = string.Format(ModelCacheEventConsumer.PRODUCT_DEFAULTPICTURE_MODEL_KEY, product.Id, pictureSize, true, _workContext.WorkingLanguage.Id, _webHelper.IsCurrentConnectionSecured(), _storeContext.CurrentStore.Id);
+                    model.DefaultPictureModel = _cacheManager.Get(defaultProductPictureCacheKey, () =>
+                    {
+                        var picture = _pictureService.GetPicturesByProductId(product.Id, 1).FirstOrDefault();
+                        var pictureModel = new PictureModel
+                        {
+                            ImageUrl = _pictureService.GetPictureUrl(picture, pictureSize),
+                            FullSizeImageUrl = _pictureService.GetPictureUrl(picture)
+                        };
+                        //"title" attribute
+                        pictureModel.Title = (picture != null && !string.IsNullOrEmpty(picture.TitleAttribute)) ?
+                            picture.TitleAttribute :
+                            string.Format(_localizationService.GetResource("Media.Product.ImageLinkTitleFormat"), model.Name);
+                        //"alt" attribute
+                        pictureModel.AlternateText = (picture != null && !string.IsNullOrEmpty(picture.AltAttribute)) ?
+                            picture.AltAttribute :
+                            string.Format(_localizationService.GetResource("Media.Product.ImageAlternateTextFormat"), model.Name);
+
+                        return pictureModel;
+                    });
+
+                    #endregion
+                }
+
+                //specs
+                if (prepareSpecificationAttributes)
+                {
+                    model.SpecificationAttributeModels = PrepareProductSpecificationModel(product);
+                }
+
+                //reviews
+                model.ReviewOverviewModel = PrepareProductReviewOverviewModel(product);
+
+                models.Add(model);
+            }
+            return models;
+        }
+
+        public virtual IList<ProductSpecificationModel> PrepareProductSpecificationModel(Product product)
+        {
+            if (product == null)
+                throw new ArgumentNullException("product");
+
+            string cacheKey = string.Format(ModelCacheEventConsumer.PRODUCT_SPECS_MODEL_KEY, product.Id, _workContext.WorkingLanguage.Id);
+            return _cacheManager.Get(cacheKey, () =>
+                _specificationAttributeService.GetProductSpecificationAttributes(product.Id, 0, null, true)
+                .Select(psa =>
+                {
+                    var m = new ProductSpecificationModel
+                    {
+                        SpecificationAttributeId = psa.SpecificationAttributeOption.SpecificationAttributeId,
+                        SpecificationAttributeName = psa.SpecificationAttributeOption.SpecificationAttribute.GetLocalized(x => x.Name),
+                        ColorSquaresRgb = psa.SpecificationAttributeOption.ColorSquaresRgb
+                    };
+
+                    switch (psa.AttributeType)
+                    {
+                        case SpecificationAttributeType.Option:
+                            m.ValueRaw = HttpUtility.HtmlEncode(psa.SpecificationAttributeOption.GetLocalized(x => x.Name));
+                            break;
+                        case SpecificationAttributeType.CustomText:
+                            m.ValueRaw = HttpUtility.HtmlEncode(psa.CustomValue);
+                            break;
+                        case SpecificationAttributeType.CustomHtmlText:
+                            m.ValueRaw = psa.CustomValue;
+                            break;
+                        case SpecificationAttributeType.Hyperlink:
+                            m.ValueRaw = string.Format("<a href='{0}' target='_blank'>{0}</a>", psa.CustomValue);
+                            break;
+                        default:
+                            break;
+                    }
+                    return m;
+                }).ToList()
+            );
+        }
+
+        #endregion
+
+        #region Categories
+
+        public virtual CategoryModel PrepareCategoryModel(Category category, CatalogPagingFilteringModel command)
+        {
+            if (category == null)
+                throw new ArgumentNullException("category");
+
+            var model = category.ToModel();
+            
+
+            //sorting
+            PrepareSortingOptions(model.PagingFilteringContext, command);
+            //view mode
+            PrepareViewModes(model.PagingFilteringContext, command);
+            //page size
+            PreparePageSizeOptions(model.PagingFilteringContext, command,
+                category.AllowCustomersToSelectPageSize, 
+                category.PageSizeOptions, 
+                category.PageSize);
+
+            //price ranges
+            model.PagingFilteringContext.PriceRangeFilter.LoadPriceRangeFilters(category.PriceRanges, _webHelper, _priceFormatter);
+            var selectedPriceRange = model.PagingFilteringContext.PriceRangeFilter.GetSelectedPriceRange(_webHelper, category.PriceRanges);
+            decimal? minPriceConverted = null;
+            decimal? maxPriceConverted = null;
+            if (selectedPriceRange != null)
+            {
+                if (selectedPriceRange.From.HasValue)
+                    minPriceConverted = _currencyService.ConvertToPrimaryStoreCurrency(selectedPriceRange.From.Value, _workContext.WorkingCurrency);
+
+                if (selectedPriceRange.To.HasValue)
+                    maxPriceConverted = _currencyService.ConvertToPrimaryStoreCurrency(selectedPriceRange.To.Value, _workContext.WorkingCurrency);
+            }
+
+
+            //category breadcrumb
+            if (_catalogSettings.CategoryBreadcrumbEnabled)
+            {
+                model.DisplayCategoryBreadcrumb = true;
+
+                string breadcrumbCacheKey = string.Format(ModelCacheEventConsumer.CATEGORY_BREADCRUMB_KEY,
+                    category.Id,
+                    string.Join(",", _workContext.CurrentCustomer.GetCustomerRoleIds()),
+                    _storeContext.CurrentStore.Id,
+                    _workContext.WorkingLanguage.Id);
+                model.CategoryBreadcrumb = _cacheManager.Get(breadcrumbCacheKey, () =>
+                    category
+                    .GetCategoryBreadCrumb(_categoryService, _aclService, _storeMappingService)
+                    .Select(catBr => new CategoryModel
+                    {
+                        Id = catBr.Id,
+                        Name = catBr.GetLocalized(x => x.Name),
+                        SeName = catBr.GetSeName()
+                    })
+                    .ToList()
+                );
+            }
+
+
+            
+            var pictureSize = _mediaSettings.CategoryThumbPictureSize;
+
+            //subcategories
+            string subCategoriesCacheKey = string.Format(ModelCacheEventConsumer.CATEGORY_SUBCATEGORIES_KEY,
+                category.Id,
+                pictureSize,
+                string.Join(",", _workContext.CurrentCustomer.GetCustomerRoleIds()),
+                _storeContext.CurrentStore.Id,
+                _workContext.WorkingLanguage.Id,
+                _webHelper.IsCurrentConnectionSecured());
+            model.SubCategories = _cacheManager.Get(subCategoriesCacheKey, () =>
+                _categoryService.GetAllCategoriesByParentCategoryId(category.Id)
+                .Select(x =>
+                {
+                    var subCatModel = new CategoryModel.SubCategoryModel
+                    {
+                        Id = x.Id,
+                        Name = x.GetLocalized(y => y.Name),
+                        SeName = x.GetSeName(),
+                        Description = x.GetLocalized(y => y.Description)
+                    };
+
+                    //prepare picture model
+                    var categoryPictureCacheKey = string.Format(ModelCacheEventConsumer.CATEGORY_PICTURE_MODEL_KEY, x.Id, pictureSize, true, _workContext.WorkingLanguage.Id, _webHelper.IsCurrentConnectionSecured(), _storeContext.CurrentStore.Id);
+                    subCatModel.PictureModel = _cacheManager.Get(categoryPictureCacheKey, () =>
+                    {
+                        var picture = _pictureService.GetPictureById(x.PictureId);
+                        var pictureModel = new PictureModel
+                        {
+                            FullSizeImageUrl = _pictureService.GetPictureUrl(picture),
+                            ImageUrl = _pictureService.GetPictureUrl(picture, pictureSize),
+                            Title = string.Format(_localizationService.GetResource("Media.Category.ImageLinkTitleFormat"), subCatModel.Name),
+                            AlternateText = string.Format(_localizationService.GetResource("Media.Category.ImageAlternateTextFormat"), subCatModel.Name)
+                        };
+                        return pictureModel;
+                    });
+
+                    return subCatModel;
+                })
+                .ToList()
+            );
+
+
+
+
+            //featured products
+            if (!_catalogSettings.IgnoreFeaturedProducts)
+            {
+                //We cache a value indicating whether we have featured products
+                IPagedList<Product> featuredProducts = null;
+                string cacheKey = string.Format(ModelCacheEventConsumer.CATEGORY_HAS_FEATURED_PRODUCTS_KEY, category.Id,
+                    string.Join(",", _workContext.CurrentCustomer.GetCustomerRoleIds()), _storeContext.CurrentStore.Id);
+                var hasFeaturedProductsCache = _cacheManager.Get<bool?>(cacheKey);
+                if (!hasFeaturedProductsCache.HasValue)
+                {
+                    //no value in the cache yet
+                    //let's load products and cache the result (true/false)
+                    featuredProducts = _productService.SearchProducts(
+                       categoryIds: new List<int> { category.Id },
+                       storeId: _storeContext.CurrentStore.Id,
+                       visibleIndividuallyOnly: true,
+                       featuredProducts: true);
+                    hasFeaturedProductsCache = featuredProducts.TotalCount > 0;
+                    _cacheManager.Set(cacheKey, hasFeaturedProductsCache, 60);
+                }
+                if (hasFeaturedProductsCache.Value && featuredProducts == null)
+                {
+                    //cache indicates that the category has featured products
+                    //let's load them
+                    featuredProducts = _productService.SearchProducts(
+                       categoryIds: new List<int> { category.Id },
+                       storeId: _storeContext.CurrentStore.Id,
+                       visibleIndividuallyOnly: true,
+                       featuredProducts: true);
+                }
+                if (featuredProducts != null)
+                {
+                    model.FeaturedProducts = PrepareProductOverviewModels(featuredProducts).ToList();
+                }
+            }
+
+
+            var categoryIds = new List<int>();
+            categoryIds.Add(category.Id);
+            if (_catalogSettings.ShowProductsFromSubcategories)
+            {
+                //include subcategories
+                categoryIds.AddRange(GetChildCategoryIds(category.Id));
+            }
+            //products
+            IList<int> alreadyFilteredSpecOptionIds = model.PagingFilteringContext.SpecificationFilter.GetAlreadyFilteredSpecOptionIds(_webHelper);
+            IList<int> filterableSpecificationAttributeOptionIds;
+            var products = _productService.SearchProducts(out filterableSpecificationAttributeOptionIds,
+                true,
+                categoryIds: categoryIds,
+                storeId: _storeContext.CurrentStore.Id,
+                visibleIndividuallyOnly: true,
+                featuredProducts:_catalogSettings.IncludeFeaturedProductsInNormalLists ? null : (bool?)false,
+                priceMin:minPriceConverted, 
+                priceMax:maxPriceConverted,
+                filteredSpecs: alreadyFilteredSpecOptionIds,
+                orderBy: (ProductSortingEnum)command.OrderBy,
+                pageIndex: command.PageNumber - 1,
+                pageSize: command.PageSize);
+            model.Products = PrepareProductOverviewModels(products).ToList();
+
+            model.PagingFilteringContext.LoadPagedList(products);
+
+            //specs
+            model.PagingFilteringContext.SpecificationFilter.PrepareSpecsFilters(alreadyFilteredSpecOptionIds,
+                filterableSpecificationAttributeOptionIds != null ? filterableSpecificationAttributeOptionIds.ToArray() : null, 
+                _specificationAttributeService, 
+                _webHelper, 
+                _workContext,
+                _cacheManager);
+            
+            return model;
+        }
+
+        public virtual string PrepareCategoryTemplateViewPath(int templateId)
+        {
+            var templateCacheKey = string.Format(ModelCacheEventConsumer.CATEGORY_TEMPLATE_MODEL_KEY, templateId);
+            var templateViewPath = _cacheManager.Get(templateCacheKey, () =>
+            {
+                var template = _categoryTemplateService.GetCategoryTemplateById(templateId);
+                if (template == null)
+                    template = _categoryTemplateService.GetAllCategoryTemplates().FirstOrDefault();
+                if (template == null)
+                    throw new Exception("No default template could be loaded");
+                return template.ViewPath;
+            });
+
+            return templateViewPath;
+        }
+
+        public virtual CategoryNavigationModel PrepareCategoryNavigationModel(int currentCategoryId, int currentProductId)
+        {
+            //get active category
+            int activeCategoryId = 0;
+            if (currentCategoryId > 0)
+            {
+                //category details page
+                activeCategoryId = currentCategoryId;
+            }
+            else if (currentProductId > 0)
+            {
+                //product details page
+                var productCategories = _categoryService.GetProductCategoriesByProductId(currentProductId);
+                if (productCategories.Any())
+                    activeCategoryId = productCategories[0].CategoryId;
+            }
+
+            string cacheKey = string.Format(ModelCacheEventConsumer.CATEGORY_NAVIGATION_MODEL_KEY, 
+                _workContext.WorkingLanguage.Id,
+                string.Join(",", _workContext.CurrentCustomer.GetCustomerRoleIds()), 
+                _storeContext.CurrentStore.Id);
+            var cachedModel = _cacheManager.Get(cacheKey, () => PrepareCategorySimpleModels(0).ToList());
+
+            var model = new CategoryNavigationModel
+            {
+                CurrentCategoryId = activeCategoryId,
+                Categories = cachedModel
+            };
+
+            return model;
+        }
+
+        public virtual TopMenuModel PrepareTopMenuModel()
+        {
+            //categories
+            string categoryCacheKey = string.Format(ModelCacheEventConsumer.CATEGORY_MENU_MODEL_KEY,
+                _workContext.WorkingLanguage.Id,
+                string.Join(",", _workContext.CurrentCustomer.GetCustomerRoleIds()), 
+                _storeContext.CurrentStore.Id);
+            var cachedCategoriesModel = _cacheManager.Get(categoryCacheKey, () => PrepareCategorySimpleModels(0));
+
+            //top menu topics
+            string topicCacheKey = string.Format(ModelCacheEventConsumer.TOPIC_TOP_MENU_MODEL_KEY, 
+                _workContext.WorkingLanguage.Id,
+                _storeContext.CurrentStore.Id,
+                string.Join(",", _workContext.CurrentCustomer.GetCustomerRoleIds()));
+            var cachedTopicModel = _cacheManager.Get(topicCacheKey, () =>
+                _topicService.GetAllTopics(_storeContext.CurrentStore.Id)
+                .Where(t => t.IncludeInTopMenu)
+                .Select(t => new TopMenuModel.TopMenuTopicModel
+                {
+                    Id = t.Id,
+                    Name = t.GetLocalized(x => x.Title),
+                    SeName = t.GetSeName()
+                })
+                .ToList()
+            );
+            var model = new TopMenuModel
+            {
+                Categories = cachedCategoriesModel,
+                Topics = cachedTopicModel,
+                NewProductsEnabled = _catalogSettings.NewProductsEnabled,
+                BlogEnabled = _blogSettings.Enabled,
+                ForumEnabled = _forumSettings.ForumsEnabled
+            };
+            return model;
+        }
+        
+        public virtual List<CategoryModel> PrepareHomepageCategoryModels()
+        {
+            var pictureSize = _mediaSettings.CategoryThumbPictureSize;
+
+            string categoriesCacheKey = string.Format(ModelCacheEventConsumer.CATEGORY_HOMEPAGE_KEY,
+                string.Join(",", _workContext.CurrentCustomer.GetCustomerRoleIds()), 
+                pictureSize,
+                _storeContext.CurrentStore.Id,
+                _workContext.WorkingLanguage.Id, 
+                _webHelper.IsCurrentConnectionSecured());
+
+            var model = _cacheManager.Get(categoriesCacheKey, () =>
+                _categoryService.GetAllCategoriesDisplayedOnHomePage()
+                .Select(x =>
+                {
+                    var catModel = x.ToModel();
+
+                    //prepare picture model
+                    var categoryPictureCacheKey = string.Format(ModelCacheEventConsumer.CATEGORY_PICTURE_MODEL_KEY, x.Id, pictureSize, true, _workContext.WorkingLanguage.Id, _webHelper.IsCurrentConnectionSecured(), _storeContext.CurrentStore.Id);
+                    catModel.PictureModel = _cacheManager.Get(categoryPictureCacheKey, () =>
+                    {
+                        var picture = _pictureService.GetPictureById(x.PictureId);
+                        var pictureModel = new PictureModel
+                        {
+                            FullSizeImageUrl = _pictureService.GetPictureUrl(picture),
+                            ImageUrl = _pictureService.GetPictureUrl(picture, pictureSize),
+                            Title = string.Format(_localizationService.GetResource("Media.Category.ImageLinkTitleFormat"), catModel.Name),
+                            AlternateText = string.Format(_localizationService.GetResource("Media.Category.ImageAlternateTextFormat"), catModel.Name)
+                        };
+                        return pictureModel;
+                    });
+
+                    return catModel;
+                })
+                .ToList()
+            );
+
+            return model;
+        }
+
+        #endregion
+
+        #region Manufacturers
+
+        public virtual ManufacturerModel PrepareManufacturerModel(Manufacturer manufacturer, CatalogPagingFilteringModel command)
+        {
+            if (manufacturer == null)
+                throw new ArgumentNullException("manufacturer");
+
+            var model = manufacturer.ToModel();
+
+
+
+            //sorting
+            PrepareSortingOptions(model.PagingFilteringContext, command);
+            //view mode
+            PrepareViewModes(model.PagingFilteringContext, command);
+            //page size
+            PreparePageSizeOptions(model.PagingFilteringContext, command,
+                manufacturer.AllowCustomersToSelectPageSize,
+                manufacturer.PageSizeOptions,
+                manufacturer.PageSize);
+
+
+            //price ranges
+            model.PagingFilteringContext.PriceRangeFilter.LoadPriceRangeFilters(manufacturer.PriceRanges, _webHelper, _priceFormatter);
+            var selectedPriceRange = model.PagingFilteringContext.PriceRangeFilter.GetSelectedPriceRange(_webHelper, manufacturer.PriceRanges);
+            decimal? minPriceConverted = null;
+            decimal? maxPriceConverted = null;
+            if (selectedPriceRange != null)
+            {
+                if (selectedPriceRange.From.HasValue)
+                    minPriceConverted = _currencyService.ConvertToPrimaryStoreCurrency(selectedPriceRange.From.Value, _workContext.WorkingCurrency);
+
+                if (selectedPriceRange.To.HasValue)
+                    maxPriceConverted = _currencyService.ConvertToPrimaryStoreCurrency(selectedPriceRange.To.Value, _workContext.WorkingCurrency);
+            }
+
+
+
+            //featured products
+            if (!_catalogSettings.IgnoreFeaturedProducts)
+            {
+                IPagedList<Product> featuredProducts = null;
+
+                //We cache a value indicating whether we have featured products
+                string cacheKey = string.Format(ModelCacheEventConsumer.MANUFACTURER_HAS_FEATURED_PRODUCTS_KEY, 
+                    manufacturer.Id,
+                    string.Join(",", _workContext.CurrentCustomer.GetCustomerRoleIds()),
+                    _storeContext.CurrentStore.Id);
+                var hasFeaturedProductsCache = _cacheManager.Get<bool?>(cacheKey);
+                if (!hasFeaturedProductsCache.HasValue)
+                {
+                    //no value in the cache yet
+                    //let's load products and cache the result (true/false)
+                    featuredProducts = _productService.SearchProducts(
+                       manufacturerId: manufacturer.Id,
+                       storeId: _storeContext.CurrentStore.Id,
+                       visibleIndividuallyOnly: true,
+                       featuredProducts: true);
+                    hasFeaturedProductsCache = featuredProducts.TotalCount > 0;
+                    _cacheManager.Set(cacheKey, hasFeaturedProductsCache, 60);
+                }
+                if (hasFeaturedProductsCache.Value && featuredProducts == null)
+                {
+                    //cache indicates that the manufacturer has featured products
+                    //let's load them
+                    featuredProducts = _productService.SearchProducts(
+                       manufacturerId: manufacturer.Id,
+                       storeId: _storeContext.CurrentStore.Id,
+                       visibleIndividuallyOnly: true,
+                       featuredProducts: true);
+                }
+                if (featuredProducts != null)
+                {
+                    model.FeaturedProducts = PrepareProductOverviewModels(featuredProducts).ToList();
+                }
+            }
+
+
+
+            //products
+            IList<int> filterableSpecificationAttributeOptionIds;
+            var products = _productService.SearchProducts(out filterableSpecificationAttributeOptionIds, true,
+                manufacturerId: manufacturer.Id,
+                storeId: _storeContext.CurrentStore.Id,
+                visibleIndividuallyOnly: true,
+                featuredProducts: _catalogSettings.IncludeFeaturedProductsInNormalLists ? null : (bool?)false,
+                priceMin: minPriceConverted, 
+                priceMax: maxPriceConverted,
+                orderBy: (ProductSortingEnum)command.OrderBy,
+                pageIndex: command.PageNumber - 1,
+                pageSize: command.PageSize);
+            model.Products = PrepareProductOverviewModels(products).ToList();
+
+            model.PagingFilteringContext.LoadPagedList(products);
+
+            return model;
+        }
+
+        public virtual string PrepareManufacturerTemplateViewPath(int templateId)
+        {
+            var templateCacheKey = string.Format(ModelCacheEventConsumer.MANUFACTURER_TEMPLATE_MODEL_KEY, templateId);
+            var templateViewPath = _cacheManager.Get(templateCacheKey, () =>
+            {
+                var template = _manufacturerTemplateService.GetManufacturerTemplateById(templateId);
+                if (template == null)
+                    template = _manufacturerTemplateService.GetAllManufacturerTemplates().FirstOrDefault();
+                if (template == null)
+                    throw new Exception("No default template could be loaded");
+                return template.ViewPath;
+            });
+
+            return templateViewPath;
+        }
+
+        public virtual List<ManufacturerModel> PrepareManufacturerAllModels()
+        {
+            var model = new List<ManufacturerModel>();
+            var manufacturers = _manufacturerService.GetAllManufacturers(storeId: _storeContext.CurrentStore.Id);
+            foreach (var manufacturer in manufacturers)
+            {
+                var modelMan = manufacturer.ToModel();
+                
+                //prepare picture model
+                int pictureSize = _mediaSettings.ManufacturerThumbPictureSize;
+                var manufacturerPictureCacheKey = string.Format(ModelCacheEventConsumer.MANUFACTURER_PICTURE_MODEL_KEY, manufacturer.Id, pictureSize, true, _workContext.WorkingLanguage.Id, _webHelper.IsCurrentConnectionSecured(), _storeContext.CurrentStore.Id);
+                modelMan.PictureModel = _cacheManager.Get(manufacturerPictureCacheKey, () =>
+                {
+                    var picture = _pictureService.GetPictureById(manufacturer.PictureId);
+                    var pictureModel = new PictureModel
+                    {
+                        FullSizeImageUrl = _pictureService.GetPictureUrl(picture),
+                        ImageUrl = _pictureService.GetPictureUrl(picture, pictureSize),
+                        Title = string.Format(_localizationService.GetResource("Media.Manufacturer.ImageLinkTitleFormat"), modelMan.Name),
+                        AlternateText = string.Format(_localizationService.GetResource("Media.Manufacturer.ImageAlternateTextFormat"), modelMan.Name)
+                    };
+                    return pictureModel;
+                });
+                model.Add(modelMan);
+            }
+
+            return model;
+        }
+
+        public virtual ManufacturerNavigationModel PrepareManufacturerNavigationModel(int currentManufacturerId)
+        {
+            string cacheKey = string.Format(ModelCacheEventConsumer.MANUFACTURER_NAVIGATION_MODEL_KEY, 
+                currentManufacturerId, 
+                _workContext.WorkingLanguage.Id,
+                string.Join(",", _workContext.CurrentCustomer.GetCustomerRoleIds()),
+                _storeContext.CurrentStore.Id);
+            var cacheModel = _cacheManager.Get(cacheKey, () =>
+            {
+                var currentManufacturer = _manufacturerService.GetManufacturerById(currentManufacturerId);
+
+                var manufacturers = _manufacturerService.GetAllManufacturers(storeId: _storeContext.CurrentStore.Id,
+                    pageSize: _catalogSettings.ManufacturersBlockItemsToDisplay);
+                var model = new ManufacturerNavigationModel
+                {
+                    TotalManufacturers = manufacturers.TotalCount
+                };
+
+                foreach (var manufacturer in manufacturers)
+                {
+                    var modelMan = new ManufacturerBriefInfoModel
+                    {
+                        Id = manufacturer.Id,
+                        Name = manufacturer.GetLocalized(x => x.Name),
+                        SeName = manufacturer.GetSeName(),
+                        IsActive = currentManufacturer != null && currentManufacturer.Id == manufacturer.Id,
+                    };
+                    model.Manufacturers.Add(modelMan);
+                }
+                return model;
+            });
+            
+            
+            return cacheModel;
+        }
+
+        #endregion
+
+        #region Vendors
+
+        public virtual VendorModel PrepareVendorModel(Vendor vendor, CatalogPagingFilteringModel command)
+        {
+            if (vendor == null)
+                throw new ArgumentNullException("vendor");
+
+            var model = new VendorModel
+            {
+                Id = vendor.Id,
+                Name = vendor.GetLocalized(x => x.Name),
+                Description = vendor.GetLocalized(x => x.Description),
+                MetaKeywords = vendor.GetLocalized(x => x.MetaKeywords),
+                MetaDescription = vendor.GetLocalized(x => x.MetaDescription),
+                MetaTitle = vendor.GetLocalized(x => x.MetaTitle),
+                SeName = vendor.GetSeName(),
+                AllowCustomersToContactVendors = _vendorSettings.AllowCustomersToContactVendors
+            };
+
+
+
+            //sorting
+            PrepareSortingOptions(model.PagingFilteringContext, command);
+            //view mode
+            PrepareViewModes(model.PagingFilteringContext, command);
+            //page size
+            PreparePageSizeOptions(model.PagingFilteringContext, command,
+                vendor.AllowCustomersToSelectPageSize,
+                vendor.PageSizeOptions,
+                vendor.PageSize);
+
+            //products
+            IList<int> filterableSpecificationAttributeOptionIds;
+            var products = _productService.SearchProducts(out filterableSpecificationAttributeOptionIds, true,
+                vendorId: vendor.Id,
+                storeId: _storeContext.CurrentStore.Id,
+                visibleIndividuallyOnly: true,
+                orderBy: (ProductSortingEnum)command.OrderBy,
+                pageIndex: command.PageNumber - 1,
+                pageSize: command.PageSize);
+            model.Products = PrepareProductOverviewModels(products).ToList();
+
+            model.PagingFilteringContext.LoadPagedList(products);
+
+            return model;
+        }
+
+        public virtual List<VendorModel> PrepareVendorAllModels()
+        {
+            var model = new List<VendorModel>();
+            var vendors = _vendorService.GetAllVendors();
+            foreach (var vendor in vendors)
+            {
+                var vendorModel = new VendorModel
+                {
+                    Id = vendor.Id,
+                    Name = vendor.GetLocalized(x => x.Name),
+                    Description = vendor.GetLocalized(x => x.Description),
+                    MetaKeywords = vendor.GetLocalized(x => x.MetaKeywords),
+                    MetaDescription = vendor.GetLocalized(x => x.MetaDescription),
+                    MetaTitle = vendor.GetLocalized(x => x.MetaTitle),
+                    SeName = vendor.GetSeName(),
+                    AllowCustomersToContactVendors = _vendorSettings.AllowCustomersToContactVendors
+                };
+                //prepare picture model
+                int pictureSize = _mediaSettings.VendorThumbPictureSize;
+                var pictureCacheKey = string.Format(ModelCacheEventConsumer.VENDOR_PICTURE_MODEL_KEY, vendor.Id, pictureSize, true, _workContext.WorkingLanguage.Id, _webHelper.IsCurrentConnectionSecured(), _storeContext.CurrentStore.Id);
+                vendorModel.PictureModel = _cacheManager.Get(pictureCacheKey, () =>
+                {
+                    var picture = _pictureService.GetPictureById(vendor.PictureId);
+                    var pictureModel = new PictureModel
+                    {
+                        FullSizeImageUrl = _pictureService.GetPictureUrl(picture),
+                        ImageUrl = _pictureService.GetPictureUrl(picture, pictureSize),
+                        Title = string.Format(_localizationService.GetResource("Media.Vendor.ImageLinkTitleFormat"), vendorModel.Name),
+                        AlternateText = string.Format(_localizationService.GetResource("Media.Vendor.ImageAlternateTextFormat"), vendorModel.Name)
+                    };
+                    return pictureModel;
+                });
+                model.Add(vendorModel);
+            }
+
+            return model;
+        }
+
+        public virtual VendorNavigationModel PrepareVendorNavigationModel()
+        {
+            string cacheKey = ModelCacheEventConsumer.VENDOR_NAVIGATION_MODEL_KEY;
+            var cacheModel = _cacheManager.Get(cacheKey, () =>
+            {
+                var vendors = _vendorService.GetAllVendors(pageSize: _vendorSettings.VendorsBlockItemsToDisplay);
+                var model = new VendorNavigationModel
+                {
+                    TotalVendors = vendors.TotalCount
+                };
+
+                foreach (var vendor in vendors)
+                {
+                    model.Vendors.Add(new VendorBriefInfoModel
+                    {
+                        Id = vendor.Id,
+                        Name = vendor.GetLocalized(x => x.Name),
+                        SeName = vendor.GetSeName(),
+                    });
+                }
+                return model;
+            });
+            
+            return cacheModel;
+        }
+
+        #endregion
+
+        #region Product tags
+        
+        public virtual PopularProductTagsModel PreparePopularProductTagsModel()
+        {
+            var cacheKey = string.Format(ModelCacheEventConsumer.PRODUCTTAG_POPULAR_MODEL_KEY, _workContext.WorkingLanguage.Id, _storeContext.CurrentStore.Id);
+            var cacheModel = _cacheManager.Get(cacheKey, () =>
+            {
+                var model = new PopularProductTagsModel();
+
+                //get all tags
+                var allTags = _productTagService
+                    .GetAllProductTags()
+                    //filter by current store
+                    .Where(x => _productTagService.GetProductCount(x.Id, _storeContext.CurrentStore.Id) > 0)
+                    //order by product count
+                    .OrderByDescending(x => _productTagService.GetProductCount(x.Id, _storeContext.CurrentStore.Id))
+                    .ToList();
+
+                var tags = allTags
+                    .Take(_catalogSettings.NumberOfProductTags)
+                    .ToList();
+                //sorting
+                tags = tags.OrderBy(x => x.GetLocalized(y => y.Name)).ToList();
+
+                model.TotalTags = allTags.Count;
+
+                foreach (var tag in tags)
+                    model.Tags.Add(new ProductTagModel
+                    {
+                        Id = tag.Id,
+                        Name = tag.GetLocalized(y => y.Name),
+                        SeName = tag.GetSeName(),
+                        ProductCount = _productTagService.GetProductCount(tag.Id, _storeContext.CurrentStore.Id)
+                    });
+                return model;
+            });
+
+            return cacheModel;
+        }
+
+        public virtual ProductsByTagModel PrepareProductsByTagModel(ProductTag productTag, CatalogPagingFilteringModel command)
+        {
+            if (productTag == null)
+                throw new ArgumentNullException("productTag");
+
+            var model = new ProductsByTagModel
+            {
+                Id = productTag.Id,
+                TagName = productTag.GetLocalized(y => y.Name),
+                TagSeName = productTag.GetSeName()
+            };
+
+
+            //sorting
+            PrepareSortingOptions(model.PagingFilteringContext, command);
+            //view mode
+            PrepareViewModes(model.PagingFilteringContext, command);
+            //page size
+            PreparePageSizeOptions(model.PagingFilteringContext, command,
+                _catalogSettings.ProductsByTagAllowCustomersToSelectPageSize,
+                _catalogSettings.ProductsByTagPageSizeOptions,
+                _catalogSettings.ProductsByTagPageSize);
+
+
+            //products
+            var products = _productService.SearchProducts(
+                storeId: _storeContext.CurrentStore.Id,
+                productTagId: productTag.Id,
+                visibleIndividuallyOnly: true,
+                orderBy: (ProductSortingEnum)command.OrderBy,
+                pageIndex: command.PageNumber - 1,
+                pageSize: command.PageSize);
+            model.Products = PrepareProductOverviewModels(products).ToList();
+
+            model.PagingFilteringContext.LoadPagedList(products);
+            return model;
+        }
+        
+
+        public virtual PopularProductTagsModel PrepareProductTagsAllModel()
+        {
+            var model = new PopularProductTagsModel();
+            model.Tags = _productTagService
+                .GetAllProductTags()
+                //filter by current store
+                .Where(x => _productTagService.GetProductCount(x.Id, _storeContext.CurrentStore.Id) > 0)
+                //sort by name
+                .OrderBy(x => x.GetLocalized(y => y.Name))
+                .Select(x =>
+                {
+                    var ptModel = new ProductTagModel
+                    {
+                        Id = x.Id,
+                        Name = x.GetLocalized(y => y.Name),
+                        SeName = x.GetSeName(),
+                        ProductCount = _productTagService.GetProductCount(x.Id, _storeContext.CurrentStore.Id)
+                    };
+                    return ptModel;
+                })
+                .ToList();
+            return model;
+        }
+
+        #endregion
+
+        #region Searching
+
+        public virtual SearchModel PrepareSearchModel(SearchModel model, CatalogPagingFilteringModel command)
+        {
+            if (model == null)
+                throw new ArgumentNullException("model");
+
+            var searchTerms = model.q;
+            if (searchTerms == null)
+                searchTerms = "";
+            searchTerms = searchTerms.Trim();
+
+
+            //sorting
+            PrepareSortingOptions(model.PagingFilteringContext, command);
+            //view mode
+            PrepareViewModes(model.PagingFilteringContext, command);
+            //page size
+            PreparePageSizeOptions(model.PagingFilteringContext, command,
+                _catalogSettings.SearchPageAllowCustomersToSelectPageSize,
+                _catalogSettings.SearchPagePageSizeOptions,
+                _catalogSettings.SearchPageProductsPerPage);
+
+
+            string cacheKey = string.Format(ModelCacheEventConsumer.SEARCH_CATEGORIES_MODEL_KEY, 
+                _workContext.WorkingLanguage.Id,
+                string.Join(",", _workContext.CurrentCustomer.GetCustomerRoleIds()), 
+                _storeContext.CurrentStore.Id); 
+            var categories = _cacheManager.Get(cacheKey, () =>
+            {
+                var categoriesModel = new List<SearchModel.CategoryModel>();
+                //all categories
+                var allCategories = _categoryService.GetAllCategories(storeId: _storeContext.CurrentStore.Id);
+                foreach (var c in allCategories)
+                {
+                    //generate full category name (breadcrumb)
+                    string categoryBreadcrumb= "";
+                    var breadcrumb = c.GetCategoryBreadCrumb(allCategories, _aclService, _storeMappingService);
+                    for (int i = 0; i <= breadcrumb.Count - 1; i++)
+                    {
+                        categoryBreadcrumb += breadcrumb[i].GetLocalized(x => x.Name);
+                        if (i != breadcrumb.Count - 1)
+                            categoryBreadcrumb += " >> ";
+                    }
+                    categoriesModel.Add(new SearchModel.CategoryModel
+                    {
+                        Id = c.Id,
+                        Breadcrumb = categoryBreadcrumb
+                    });
+                }
+                return categoriesModel;
+            });
+            if (categories.Any())
+            {
+                //first empty entry
+                model.AvailableCategories.Add(new SelectListItem
+                    {
+                         Value = "0",
+                         Text = _localizationService.GetResource("Common.All")
+                    });
+                //all other categories
+                foreach (var c in categories)
+                {
+                    model.AvailableCategories.Add(new SelectListItem
+                    {
+                        Value = c.Id.ToString(),
+                        Text = c.Breadcrumb,
+                        Selected = model.cid == c.Id
+                    });
+                }
+            }
+
+            var manufacturers = _manufacturerService.GetAllManufacturers(storeId: _storeContext.CurrentStore.Id);
+            if (manufacturers.Any())
+            {
+                model.AvailableManufacturers.Add(new SelectListItem
+                {
+                    Value = "0",
+                    Text = _localizationService.GetResource("Common.All")
+                });
+                foreach (var m in manufacturers)
+                    model.AvailableManufacturers.Add(new SelectListItem
+                    {
+                        Value = m.Id.ToString(),
+                        Text = m.GetLocalized(x => x.Name),
+                        Selected = model.mid == m.Id
+                    });
+            }
+
+            model.asv = _vendorSettings.AllowSearchByVendor;
+            if (model.asv)
+            {
+                var vendors = _vendorService.GetAllVendors();
+                if (vendors.Any())
+                {
+                    model.AvailableVendors.Add(new SelectListItem
+                    {
+                        Value = "0",
+                        Text = _localizationService.GetResource("Common.All")
+                    });
+                    foreach (var vendor in vendors)
+                        model.AvailableVendors.Add(new SelectListItem
+                        {
+                            Value = vendor.Id.ToString(),
+                            Text = vendor.GetLocalized(x => x.Name),
+                            Selected = model.vid == vendor.Id
+                        });
+                }
+            }
+
+            IPagedList<Product> products = new PagedList<Product>(new List<Product>(), 0, 1);
+            var isSearchTermSpecified = false;
+            try
+            {
+                // only search if query string search keyword is set (used to avoid searching or displaying search term min length error message on /search page load)
+                isSearchTermSpecified = _httpContext.Request.Params["q"] != null;
+            }
+            catch
+            {
+                //the "A potentially dangerous Request.QueryString value was detected from the client" exception could be thrown here when some wrong char is specified (e.g. <)
+                //although we [ValidateInput(false)] attribute here we try to access "Request.Params" directly
+                //that's why we do not re-throw it
+
+                //just ensure that some search term is specified (0 length is not supported inthis case)
+                isSearchTermSpecified = !String.IsNullOrEmpty(searchTerms);
+            }
+            if (isSearchTermSpecified)
+            {
+                if (searchTerms.Length < _catalogSettings.ProductSearchTermMinimumLength)
+                {
+                    model.Warning = string.Format(_localizationService.GetResource("Search.SearchTermMinimumLengthIsNCharacters"), _catalogSettings.ProductSearchTermMinimumLength);
+                }
+                else
+                {
+                    var categoryIds = new List<int>();
+                    int manufacturerId = 0;
+                    decimal? minPriceConverted = null;
+                    decimal? maxPriceConverted = null;
+                    bool searchInDescriptions = false;
+                    int vendorId = 0;
+                    if (model.adv)
+                    {
+                        //advanced search
+                        var categoryId = model.cid;
+                        if (categoryId > 0)
+                        {
+                            categoryIds.Add(categoryId);
+                            if (model.isc)
+                            {
+                                //include subcategories
+                                categoryIds.AddRange(GetChildCategoryIds(categoryId));
+                            }
+                        }
+
+                        manufacturerId = model.mid;
+
+                        //min price
+                        if (!string.IsNullOrEmpty(model.pf))
+                        {
+                            decimal minPrice;
+                            if (decimal.TryParse(model.pf, out minPrice))
+                                minPriceConverted = _currencyService.ConvertToPrimaryStoreCurrency(minPrice, _workContext.WorkingCurrency);
+                        }
+                        //max price
+                        if (!string.IsNullOrEmpty(model.pt))
+                        {
+                            decimal maxPrice;
+                            if (decimal.TryParse(model.pt, out maxPrice))
+                                maxPriceConverted = _currencyService.ConvertToPrimaryStoreCurrency(maxPrice, _workContext.WorkingCurrency);
+                        }
+
+                        if (model.asv)
+                            vendorId = model.vid;
+
+                        searchInDescriptions = model.sid;
+                    }
+                    
+                    //var searchInProductTags = false;
+                    var searchInProductTags = searchInDescriptions;
+
+                    //products
+                    products = _productService.SearchProducts(
+                        categoryIds: categoryIds,
+                        manufacturerId: manufacturerId,
+                        storeId: _storeContext.CurrentStore.Id,
+                        visibleIndividuallyOnly: true,
+                        priceMin: minPriceConverted,
+                        priceMax: maxPriceConverted,
+                        keywords: searchTerms,
+                        searchDescriptions: searchInDescriptions,
+                        searchProductTags: searchInProductTags,
+                        languageId: _workContext.WorkingLanguage.Id,
+                        orderBy: (ProductSortingEnum)command.OrderBy,
+                        pageIndex: command.PageNumber - 1,
+                        pageSize: command.PageSize,
+                        vendorId: vendorId);
+                    model.Products = PrepareProductOverviewModels(products).ToList();
+
+                    model.NoResults = !model.Products.Any();
+
+                    //search term statistics
+                    if (!String.IsNullOrEmpty(searchTerms))
+                    {
+                        var searchTerm = _searchTermService.GetSearchTermByKeyword(searchTerms, _storeContext.CurrentStore.Id);
+                        if (searchTerm != null)
+                        {
+                            searchTerm.Count++;
+                            _searchTermService.UpdateSearchTerm(searchTerm);
+                        }
+                        else
+                        {
+                            searchTerm = new SearchTerm
+                            {
+                                Keyword = searchTerms,
+                                StoreId = _storeContext.CurrentStore.Id,
+                                Count = 1
+                            };
+                            _searchTermService.InsertSearchTerm(searchTerm);
+                        }
+                    }
+
+                    //event
+                    _eventPublisher.Publish(new ProductSearchEvent
+                    {
+                        SearchTerm = searchTerms,
+                        SearchInDescriptions = searchInDescriptions,
+                        CategoryIds = categoryIds,
+                        ManufacturerId = manufacturerId,
+                        WorkingLanguageId = _workContext.WorkingLanguage.Id,
+                        VendorId = vendorId
+                    });
+                }
+            }
+
+            model.PagingFilteringContext.LoadPagedList(products);
+            return model;
+        }
+
+        public virtual SearchBoxModel PrepareSearchBoxModel()
+        {
+            var model = new SearchBoxModel
+            {
+                AutoCompleteEnabled = _catalogSettings.ProductSearchAutoCompleteEnabled,
+                ShowProductImagesInSearchAutoComplete = _catalogSettings.ShowProductImagesInSearchAutoComplete,
+                SearchTermMinimumLength = _catalogSettings.ProductSearchTermMinimumLength
+            };
+            return model;
+        }
+
+        #endregion
+    }
+}
